Save button on customer edit.  Set date to date only UK on customer view
Set date on Customer index to UK short date but need more formatting

Added save button to Edit customer

Move view contact to another controller
