+++ Views/Contacts/ViewContact.cshtml
@@ -1,0 +1,91 @@
+﻿@model BlueMoonAdmin.Models.Contacts
+
+@{
+    ViewData["Title"] = "Contacts";
+}
+<form method="post" asp-action="ViewContact">
+    <input asp-for="Id" hidden />
+   
+    <div class="container-fluid">
+        <h1 class="mt-4">View Contact</h1>
+        <ol class="breadcrumb mb-4">
+            <li class="breadcrumb-item"><a asp-controller="Customer" asp-action="Index">Dashboard</a></li>
+            <li class="breadcrumb-item active">View Contact</li>
+        </ol>
+        <!-- #region Form -->
+        <div class="container col-12 border-bottom">
+            <div class="row">
+                <div class="col-12">
+                    <div class="form-group row">
+                        <div class="col-3">
+                            <label asp-for="FirstName"></label>
+                        </div>
+                        <div class="col-3">
+                            <label asp-for="LastName"></label>
+                        </div>
+                        <div class="col-3">
+                            <label asp-for="TelephoneNumber"></label>
+                        </div>
+                        <div class="col-3">
+                            <label asp-for="MobileNumber"></label>
+                        </div>
+                    </div>
+                    <div class="form-group row">
+                        <div class="col-3">
+                            <input asp-for="FirstName" class="form-control" readonly/>
+                        </div>
+                        <div class="col-3">
+                            <input asp-for="LastName" class="form-control" readonly/>
+                        </div>
+                        <div class="col-3">
+                            <input asp-for="TelephoneNumber" class="form-control" readonly>
+                        </div>
+                        <div class="col-3">
+                            <input asp-for="MobileNumber" class="form-control" readonly>
+                        </div>
+                    </div>
+                    <div class="form-group row">
+                        <div class="col-3">
+                            <label asp-for="Email"></label>
+                        </div>
+                        <div class="col-3">
+                            <label asp-for="PrimaryContact"></label>
+                        </div>
+                    </div>
+                    <div class="form-group row">
+                        <div class="col-3">
+                            <input asp-for="Email" class="form-control" readonly/>
+                        </div>
+                        <div class="col-3">
+                            <input asp-for="PrimaryContact" class="form-control" disabled />
+                        </div>                        
+                    </div>
+                    <div class="form-group row">
+                        <div class="col-8 offset-2 row">
+                            <div class="col">
+                                <a asp-action="UpdateContact" asp-route-Id="@Model.Id" class="btn btn-info w-75">Edit</a>                              
+                            </div>
+                            <div class="col">
+                                <a asp-action="ViewCustomer" asp-route-Id="@Model.CustomerId" class="btn btn-danger w-75">Back</a>
+                            </div> 
+                        </div>
+                    </div>
+
+                </div>
+            </div>
+
+            <!-- #endregion -->
+        </div>
+    </div>
+</form>
+
+
+
+@section Scripts{
+    <script src="~/assets/demo/chart-area-demo.js"></script>
+    <script src="~/assets/demo/chart-bar-demo.js"></script>
+    <script src="https://cdn.datatables.net/1.10.20/js/jquery.dataTables.min.js" crossorigin="anonymous"></script>
+    <script src="https://cdn.datatables.net/1.10.20/js/dataTables.bootstrap4.min.js" crossorigin="anonymous"></script>
+    <script src="~/assets/demo/datatables-demo.js"></script>
+}-->
+
