+++ Views/Customers/UpdateCustomer.cshtml
@@ -132,7 +132,7 @@
                     <div class="form-group row pt-4 mb-4">
                         <div class="col-8 offset-2 row">
                             <div class="col">
-                                <a asp-action="DeleteCustomer" asp-route-Id="@Model.Id" class="btn btn-danger w-75">Delete Customer</a>
+                                <input type="submit" class="btn btn-info w-75" value="Save" />
                             </div>
                             <div class="col">
                                 <a asp-action="ViewCustomer" asp-route-Id="@Model.Id" class="btn btn-danger w-75">Back</a>
