+++ Views/Customers/ViewCustomer.cshtml
@@ -175,7 +175,7 @@
                             <td>@item.PrimaryContact</td>
                             <td>
                                 <div class="w-100 btn-group" role="group">
-                                    <a asp-controller="Customers" asp-action="ViewContact" asp-route-Id="@item.Id" class="btn btn-primary mx-1">View</a>
+                                    <a asp-controller="Contacts" asp-action="ViewContact" asp-route-Id="@item.Id" class="btn btn-primary mx-1">View</a>
                                     @*<a asp-controller="Customers" asp-action="DeleteCustomer" asp-route-Id="@customer.Id" class="btn btn-danger mx-1">Delete</a>*@
                                     <a asp-controller="Customers" asp-action="UpdateContact" asp-route-Id="@item.Id" class="btn btn-danger mx-1">Edit</a>
